use class_eval in decorators
